pyproject.toml
[project]
name = "fastapi-study"
version = "0.1.0"
description = "FastAPI study project with Clean Architecture"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "pydantic-settings",
    "google-genai",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (ordenação de imports)
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]

[tool.ruff.format]
quote-style = "double"